+++ ChildProject/utils.py
@@ -25,6 +25,15 @@ class Segment:
 
     def __repr__(self):
         return "Segment([{}, {}])".format(self.start, self.stop)
+                
+def retry_func( func : callable , excep: Exception, tries : int = 3, **kwargs):
+    for i in range(tries):
+        try:
+            func(**kwargs)
+            return
+        except excep as e:
+            if i == tries - 1:
+                raise e
 
 
 def intersect_ranges(xs, ys):